# Significantly improve tight layout performance for cartopy axes