This PR is adapted from SciTools/cartopy#1956 following our discussion there.

## Summary

This PR significantly improves the speed of `Axes.get_tightbbox()` for non-rectilinear axes (in particular, matplotlib's PolarAxes and cartopy's GeoAxes) by preventing unnecessary and expensive `get_window_extent()` computations. In the presence of complex, high-resolution artists, it can result in a **2x improvement** to the draw time when "tight layout" is enabled (since artists are effectively only drawn once instead of twice).

There may be a better way to implement this -- looking forward to everyone's thoughts.

## Details

To prevent unnecessary and expensive `get_window_extent()` computations, `Axes.get_tightbbox()` skips artists with `clip_on` set to True and whose `clip_box.extents` are equivalent to `ax.bbox.extents`. However, while all axes artists are clipped by `TransformedPatchPath(ax.patch)` by default, only artists drawn inside rectilinear projections are clipped by `ax.bbox` (see the below example).

This PR replaces `Artist._get_clipping_extent_bbox()` with `Artist._is_axes_clipped()`. Now,`Axes.get_tightbbox()` includes `ax.patch.get_window_extent()` in its computation, and skips all artists clipped by **either** `ax.bbox` or `ax.patch` (i.e., artists for which `Artist._is_axes_clipped()` returns True).

This PR also removes the computation of the intersection of `clip_box` and `clip_path`, under the assumption that the independent test of `clip_path` covers those instances, but perhaps that should be added back.

## Example

Here is an example with a cartopy `GeoAxes`:

```python
import numpy as np
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
fig, ax = plt.subplots(subplot_kw={'projection': ccrs.Robinson()})
N = 5000  # large dataset
lon = np.linspace(-180, 180, N)
lat = np.linspace(-90, 90, N)
data = np.random.rand(N, N)
m = ax.pcolormesh(lon, lat, data, transform=ccrs.PlateCarree())
print(m.get_clip_box() is None)  # returns True
%timeit fig.tight_layout()
```

Performance before:

```
638 ms ± 67.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Performance after:

```
2.37 ms ± 60.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

And of course the performance difference is larger the larger the dataset.

## Checklist

<!-- Please mark any checkboxes that do not apply to this PR as [N/A]. -->
**Tests and Styling**
- [x] Has pytest style unit tests (and `pytest` passes).
- [x] Is [Flake 8](https://flake8.pycqa.org/en/latest/) compliant (install `flake8-docstrings` and run `flake8 --docstring-convention=all`).

**Documentation**
- [x] (n/a?) New features are documented, with examples if plot related.
- [x] (n/a?) New features have an entry in `doc/users/next_whats_new/` (follow instructions in README.rst there).
- [x] (n/a?) API changes documented in `doc/api/next_api_changes/` (follow instructions in README.rst there).
- [x] Documentation is sphinx and numpydoc compliant (the docs should [build](https://matplotlib.org/devel/documenting_mpl.html#building-the-docs) without error).

<!--
Thank you so much for your PR!  To help us review your contribution, please
consider the following points:

- A development guide is available at https://matplotlib.org/devdocs/devel/index.html.

- Help with git and github is available at
  https://matplotlib.org/devel/gitwash/development_workflow.html.

- Do not create the PR out of main, but out of a separate branch.

- The PR title should summarize the changes, for example "Raise ValueError on
  non-numeric input to set_xlim".  Avoid non-descriptive titles such as
  "Addresses issue #8576".

- The summary should provide at least 1-2 sentences describing the pull request
  in detail (Why is this change required?  What problem does it solve?) and
  link to any relevant issues.

- If you are contributing fixes to docstrings, please pay attention to
  http://matplotlib.org/devel/documenting_mpl.html#formatting.  In particular,
  note the difference between using single backquotes, double backquotes, and
  asterisks in the markup.

We understand that PRs can sometimes be overwhelming, especially as the
reviews start coming in.  Please let us know if the reviews are unclear or
the recommended next step seems overly demanding, if you would like help in
addressing a reviewer's comments, or if you have been waiting too long to hear
back on your PR.
-->