@lukelbd 2021-12-16
Added a simple test that `artist._is_axes_clipped()` returns `True` under default conditions / various projections. Note that all non-rectilinear axes (e.g. `PolarAxes`, `MollweideAxes`, cartopy `GeoAxes`) have `artist.get_clip_box()` set to `None`.