@lukelbd 2021-12-16
I can also see the origin of this bug: The `Artist._get_clipping_extent_bbox()` lines that I removed (previously used inside `Axes.get_tightbbox()` to skip artists) mirror the below lines in `Artist.get_tightbbox()`.

https://github.com/matplotlib/matplotlib/blob/1a74793837286b588789441c6285c174f643a5ea/lib/matplotlib/artist.py#L340-L364

While `Artist.get_tightbbox()` starts the `if` block with a valid `bbox`, `Artist._get_clipping_extent_bbox()` starts with `bbox = None`... so when it gets to Line 361, the `clip_path` is ignored, because there is no `bbox` to intersect with.