@lukelbd 2021-12-17
I've improved this PR in a force push:

* Renamed `_is_axes_clipped()` to `_fully_clipped_to_axes` (thanks @tacaswell).
* Added test lines that assert `_get_skip_layout()` is False for non-default clipping situations (re: @jklymak).
* Removed the redundant addition of `ax.patch` to the bboxes (it's included in `get_default_bbox_extra_artists()`).
* Condensed multiple return statements into a single boolean return statement (possibly cleaner?)
* Moved the line that skips artists from `Axes.get_tightbbox()` to `Axes.get_default_bbox_extra_artists()`. Note this means the public `get_default_bbox_extra_artists()` will return different values compared to previous versions.

I also made the following additional changes:

* Removed an unnecessary `bbox is not None` check in `Artist.get_tightbbox()` (this is always True).
* Removed axis instances from `get_default_bbox_extra_artists()` (this was redundant -- see `Axes.get_tightbbox()`).
* Replaced `Axes.get_tightbbox()` reference to `[xy]axis` with `ax._get_axis_list()` (consistent with removed lines).
* Always include artists that don't internally implement clipping (should fix obscure tight layout "bugs").

The latter points also resolve a second inefficiency: Currently, `Axes.get_tightbbox()` *both* 1) computes the `xaxis` and `yaxis` tight bbox using `for_layout_only=True` on these lines:

https://github.com/matplotlib/matplotlib/blob/710fce3df95e22701bd68bf6af2c8adbc9d67a79/lib/matplotlib/axes/_base.py#L4616-L4634

and 2) includes `xaxis` and `yaxis` as "extra artists" in `get_default_bbox_extra_artists()`. This meant that their bounding boxes would be effectively calculated twice if they were not "skipped" by the `_get_clipping_extent_bbox` check. For some reason, this check *was* successfully skipping `xaxis` and `yaxis` in rectilinear axes, because they have a `clip_on` set to `True` and `clip_box` set to `ax.bbox` (despite the fact that they are *not* clipped when drawn..... weird). However, this failed to skip non-rectilinear axis instances, because they only have a `clip_path` set to `ax.patch`. After this PR, axis instances are excluded from `Axes.get_default_bbox_extra_artists()`, and their extents are always calculated once.

I've also verified that, as before, `ax.patch` and `ax.spines` are still included as default "extra artists". Here's a test I used to determine the default clipping settings for various axes components:

```python
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
fig = plt.figure(figsize=(4, 2))
ax1 = fig.add_subplot(131)
ax2 = fig.add_subplot(132, projection='polar')
ax3 = fig.add_subplot(133, projection='hammer')
for ax in (ax1, ax2, ax3):
    print(ax)
    for a in (ax.patch, tuple(ax.spines.values())[0]):
        clip_box = a.get_clip_box()
        b1 = clip_box is not None and np.all(clip_box.extents == ax.bbox.extents)
        clip_path = a.get_clip_path()
        b2 = clip_path is not None and clip_path._patch is ax.patch
        print(a, '\nclip on?', a.get_clip_on(), '\naxes clip box?', b1, '\naxes clip path?', b2, '\nFULLY CLIPPED?', a._fully_clipped_to_axes())
    print()
```

Results:

```python
AxesSubplot(0.125,0.11;0.227941x0.77)
Rectangle(xy=(0, 0), width=1, height=1, angle=0)
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False
Spine
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False

PolarAxesSubplot(0.398529,0.11;0.227941x0.77)
Wedge(center=(0.5, 0.5), r=0.5, theta1=0, theta2=360, width=None)
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False
Spine
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False

HammerAxesSubplot(0.672059,0.11;0.227941x0.77)
Circle(xy=(0.5, 0.5), radius=0.5)
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False
Spine
clip on? True
axes clip box? False
axes clip path? False
FULLY CLIPPED? False
```