@jklymak 2021-12-17
This all looks good, but are we sure it has similar performance to before?  https://github.com/matplotlib/mpl-bench has benchmarking mechanisms, in particular you could use the constrained_layout and tight_layout tests?